requirements.txt
discord.py==2.2.2
discord-py-interactions
openai==0.27.0